pyproject.toml
[project]
name = "eel-hole"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask-htmx>=0.4.0",
    "flask>=3.1.0",
    "frictionless>=5.18.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.10.3",
    "duckdb>=1.1.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.7",
    "psycopg2-binary>=2.9.10",
    "authlib>=1.4.0",
    "docutils>=0.21.2",
    "structlog>=25.1.0",
    "requests>=2.32.3",
    "whoosh-reloaded>=2.7.5",
]

[tool.uv]

python-preference = "only-managed"
python-downloads = "automatic"

[dependency-groups]
dev = [
    "marimo>=0.16.0",
    "playwright>=1.55.0",
    "pre-commit>=4.5.1",
    "pytest>=8.4.2",
    "pytest-playwright>=0.7.1",
]
